BuyUIIn
Got the buy UI, and the ability to add towers in the game. Might be a
good idea to split the buys of towers and upgrades, and might also write
Upgrades as  full script

+++ Assets/Scripts/Orbiter.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Orbiter : MonoBehaviour {
     public GameObject MiddleObject;
     public GameObject Bullet;
-    public int Ring;
+    public float Ring;
     public float ShootCount;
     void Awake()
     {
